Introduced TypedArray objects.
Added the global constructors:
  Uint8Array, Int8Array, Uint8ClampedArray, Uint16Array, Int16Array,
  Uint32Array, Int32Array, Float32Array, Float64Array.

Implemented:
    - TypedArray(),
    - TypedArray(length),
    - TypedArray(buffer, [start, [end]]),
    - TypedArray(object),
    - TypedArray(typedarray),
    - TypedArray.BYTES_PER_ELEMENT.

    - %TypedArray%.prototype.constructor
    - get %TypedArray%.prototype.buffer
    - get %TypedArray%.prototype.byteLength
    - get %TypedArray%.prototype.byteOffset
    - get %TypedArray%.prototype.length
    - %TypedArray%.prototype[@@toStringTag]
    - %TypedArray%.prototype.set(array [, start])
    - %TypedArray%.prototype.slice([start, [end]])
    - %TypedArray%.prototype.toString()
    - %TypedArray%.prototype.join(separator)
    - %TypedArray%.prototype.fill().

In collaboration with Tiago Natel de Moura.

This closes #264 issue on Github.

+++ src/njs_value_conversion.h
@@ -97,7 +97,7 @@ njs_value_to_length(njs_vm_t *vm, njs_value_t *value, uint32_t *dst)
 
 
 njs_inline njs_int_t
-njs_value_to_index(njs_vm_t *vm, njs_value_t *value, uint32_t *dst)
+njs_value_to_index(njs_vm_t *vm, njs_value_t *value, uint64_t *dst)
 {
     int64_t    integer_index;
     njs_int_t  ret;
@@ -116,7 +116,7 @@ njs_value_to_index(njs_vm_t *vm, njs_value_t *value, uint32_t *dst)
             return NJS_ERROR;
         }
 
-        *dst = (uint32_t) integer_index;
+        *dst = integer_index;
     }
 
     return NJS_OK;
